requirements.txt
-i https://pypi.org/simple
amqp==5.1.1; python_version >= '3.6'
asgiref==3.7.2; python_version >= '3.7'
attrs==23.1.0; python_version >= '3.7'
autobahn==23.6.2; python_version >= '3.9'
automat==22.10.0
billiard==4.1.0; python_version >= '3.7'
celery[redis]==5.3.4; python_version >= '3.8'
cffi==1.16.0; python_version >= '3.8'
channels==4.0.0; python_version >= '3.7'
click==8.1.7; python_version >= '3.7'
click-didyoumean==0.3.0; python_full_version >= '3.6.2' and python_full_version < '4.0.0'
click-plugins==1.1.1
click-repl==0.3.0; python_version >= '3.6'
colorama==0.4.6; platform_system == 'Windows'
constantly==23.10.4; python_version >= '3.8'
cryptography==41.0.5; python_version >= '3.7'
daphne==4.0.0; python_version >= '3.7'
django==4.2.7; python_version >= '3.8'
django-cors-headers==4.3.0; python_version >= '3.8'
django-cron==0.6.0
django-decouple==2.1
djangorestframework==3.14.0; python_version >= '3.6'
djangorestframework-simplejwt==5.3.0; python_version >= '3.7'
drf-yasg==1.21.7; python_version >= '3.6'
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1; python_version >= '3.5'
kombu==5.3.2; python_version >= '3.8'
packaging==23.2; python_version >= '3.7'
prompt-toolkit==3.0.39; python_full_version >= '3.7.0'
pyasn1==0.5.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyasn1-modules==0.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pycparser==2.21
pyjwt==2.8.0; python_version >= '3.7'
pyopenssl==23.3.0
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2023.3.post1
pyyaml==6.0.1; python_version >= '3.6'
redis==4.6.0
service-identity==23.1.0
setuptools==68.2.2; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.4; python_version >= '3.5'
twisted[tls]==23.10.0; python_full_version >= '3.8.0'
twisted-iocpsupport==1.0.4; platform_system == 'Windows'
txaio==23.1.1; python_version >= '3.7'
typing-extensions==4.8.0; python_version >= '3.8'
tzdata==2023.3; python_version >= '2'
uritemplate==4.1.1; python_version >= '3.6'
vine==5.0.0; python_version >= '3.6'
wcwidth==0.2.9
zope-interface==6.1; python_version >= '3.7'
-i https://pypi.org/simple
amqp==5.1.1; python_version >= '3.6'
asgiref==3.7.2; python_version >= '3.7'
attrs==23.1.0; python_version >= '3.7'
autobahn==23.6.2; python_version >= '3.9'
automat==22.10.0
billiard==4.1.0; python_version >= '3.7'
celery[redis]==5.3.4; python_version >= '3.8'
certifi==2023.7.22; python_version >= '3.6'
cffi==1.16.0; python_version >= '3.8'
channels==4.0.0; python_version >= '3.7'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
click-didyoumean==0.3.0; python_full_version >= '3.6.2' and python_full_version < '4.0.0'
click-plugins==1.1.1
click-repl==0.3.0; python_version >= '3.6'
colorama==0.4.6; platform_system == 'Windows'
constantly==23.10.4; python_version >= '3.8'
cryptography==41.0.5; python_version >= '3.7'
daphne==4.0.0; python_version >= '3.7'
decorator==4.4.2
django==4.2.7; python_version >= '3.8'
django-cors-headers==4.3.0; python_version >= '3.8'
django-cron==0.6.0
django-decouple==2.1
djangorestframework==3.14.0; python_version >= '3.6'
djangorestframework-simplejwt==5.3.0; python_version >= '3.7'
drf-yasg==1.21.7; python_version >= '3.6'
hyperlink==21.0.0
idna==3.4; python_version >= '3.5'
imageio==2.31.6; python_version >= '3.4'
imageio-ffmpeg==0.4.9; python_version >= '3.4'
incremental==22.10.0
inflection==0.5.1; python_version >= '3.5'
kombu==5.3.2; python_version >= '3.8'
moviepy==1.0.3
numpy==1.26.1; python_version != '2.7'
packaging==23.2; python_version >= '3.7'
pillow==10.0.1; python_version >= '3.8'
proglog==0.1.10
prompt-toolkit==3.0.39; python_full_version >= '3.7.0'
pyasn1==0.5.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyasn1-modules==0.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pycparser==2.21
pyjwt==2.8.0; python_version >= '3.7'
pyopenssl==23.3.0
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2023.3.post1
pyyaml==6.0.1; python_version >= '3.6'
redis==4.6.0
requests==2.31.0; python_version >= '3.7'
service-identity==23.1.0
setuptools==68.2.2; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.4; python_version >= '3.5'
tqdm==4.66.1; python_version >= '3.7'
twisted[tls]==23.10.0; python_full_version >= '3.8.0'
twisted-iocpsupport==1.0.4; platform_system == 'Windows'
txaio==23.1.1; python_version >= '3.7'
typing-extensions==4.8.0; python_version >= '3.8'
tzdata==2023.3; python_version >= '2'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==2.0.7; python_version >= '3.7'
vine==5.0.0; python_version >= '3.6'
wcwidth==0.2.9
zope-interface==6.1; python_version >= '3.7'
